renamed definitions arrays to types
