+++ python/src/aac/lang/definitions/type.py
@@ -1,9 +1,12 @@
 """Functions for managing and interacting with definition arrays."""
 
-
 def remove_list_type_indicator(list_type: str) -> str:
     """Return the AaC type without any trailing list indicator characters '[]'."""
-    return list_type.split("[]")[0]
+    return_type = list_type
+    if is_array_type(list_type):
+        return_type = list_type.split("[]")[0]
+
+    return return_type
 
 
 def is_array_type(type: str) -> bool:
